개인 별 기록하기 데이터베이스에 연결 및 데이터 베이스에서 불러오기 함
