+++ 1.txt
@@ -1,1 +1,1 @@
-구글 지도 연결함
+개인 별 기록하기 구성함
