+++ lib/screens/home_sec1.dart
@@ -1,323 +1,37 @@
 import 'package:flutter/material.dart';
-import 'package:geolocator/geolocator.dart';
+import 'package:firebase_auth/firebase_auth.dart';
 import '../models/eco_activity.dart';
 import 'activity_auth_page.dart';
+import 'location_auth_section.dart';
+import 'activity_stats_section.dart';
+import 'activity_record_section.dart';
 
 class HomeSec1 extends StatefulWidget {
-  const HomeSec1({super.key});
+  final User user;
+
+  const HomeSec1({super.key, required this.user});
 
   @override
   State<HomeSec1> createState() => _HomeSec1State();
 }
 
 class _HomeSec1State extends State<HomeSec1> {
-  String _locationText = "지역인증을 해주세요!";
-  bool _isLocationAuthenticated = false;
-  String _localRanking = "";
-  String _contribution = "";
   late final EcoActivityData _activityData;
-  
+  bool _isLoading = true;
+
   @override
   void initState() {
     super.initState();
     _activityData = EcoActivityData();
+    _loadInitialData(widget.user);
   }
 
-  void _authenticateLocation() async {
-    try {
-      LocationPermission permission = await Geolocator.checkPermission();
-      if (permission == LocationPermission.denied) {
-        permission = await Geolocator.requestPermission();
-        if (permission == LocationPermission.denied) {
-          ScaffoldMessenger.of(context).showSnackBar(
-            const SnackBar(content: Text('위치 권한이 필요합니다')),
-          );
-          return;
-        }
-      }
-
-      if (permission == LocationPermission.deniedForever) {
-        ScaffoldMessenger.of(context).showSnackBar(
-          const SnackBar(content: Text('위치 권한이 영구적으로 거부되었습니다')),
-        );
-        return;
-      }
-
-      Position position = await Geolocator.getCurrentPosition(
-        desiredAccuracy: LocationAccuracy.high,
-      );
-
-      String location = _getLocationFromCoordinates(position.latitude, position.longitude);
-      
+  Future<void> _loadInitialData(User user) async {
+    await _activityData.loadActivitiesFromFirestore(user);
+    if (mounted) {
       setState(() {
-        _isLocationAuthenticated = true;
-        _locationText = "지역인증 완료!";
-        _localRanking = location;
-        _contribution = "상위 12%";
+        _isLoading = false;
       });
-
-    } catch (e) {
-      ScaffoldMessenger.of(context).showSnackBar(
-        SnackBar(content: Text('위치를 가져올 수 없습니다: $e')),
-      );
-    }
-  }
-
-  String _getLocationFromCoordinates(double lat, double lng) {
-    if (lat >= 37.4 && lat <= 37.6 && lng >= 126.8 && lng <= 127.0) {
-      return "금천구";
-    } else if (lat >= 37.5 && lat <= 37.6 && lng >= 127.0 && lng <= 127.1) {
-      return "강남구";
-    } else if (lat >= 37.5 && lat <= 37.6 && lng >= 126.9 && lng <= 127.0) {
-      return "서초구";
-    } else if (lat >= 37.6 && lat <= 37.7 && lng >= 126.9 && lng <= 127.1) {
-      return "성북구";
-    } else if (lat >= 37.4 && lat <= 37.5 && lng >= 126.9 && lng <= 127.0) {
-      return "영등포구";
-    } else {
-      return "서울시";
-    }
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    return Container(
-      padding: const EdgeInsets.all(20),
-      child: Column(
-        children: [
-          SizedBox(
-            height: 320,
-            child: Stack(
-              children: [
-                if (!_isLocationAuthenticated)
-                  Positioned(
-                    top: 0,
-                    left: 0,
-                    right: 0,
-                    child: Center(
-                      child: Container(
-                        padding: const EdgeInsets.symmetric(horizontal: 20, vertical: 16),
-                        decoration: BoxDecoration(
-                          color: Colors.white,
-                          borderRadius: BorderRadius.circular(25),
-                          boxShadow: [
-                            BoxShadow(
-                              color: Colors.black.withOpacity(0.1),
-                              blurRadius: 8,
-                              offset: const Offset(0, 2),
-                            ),
-                          ],
-                        ),
-                        child: Text(
-                          _locationText,
-                          style: const TextStyle(
-                            fontSize: 18,
-                            fontWeight: FontWeight.w600,
-                            color: Color(0xFF2E7D32),
-                          ),
-                        ),
-                      ),
-                    ),
-                  ),
-                if (_isLocationAuthenticated)
-                  Positioned(
-                    top: 20,
-                    left: 0,
-                    right: 0,
-                    child: Center(
-                      child: Row(
-                        mainAxisSize: MainAxisSize.min,
-                        children: [
-                          _buildRankingCard(
-                            label: "지역랭킹",
-                            value: _localRanking,
-                          ),
-                          const SizedBox(width: 16),
-                          _buildRankingCard(
-                            label: "기여도",
-                            value: _contribution,
-                          ),
-                        ],
-                      ),
-                    ),
-                  ),
-                Positioned(
-                  top: 90,
-                  left: 0,
-                  right: 0,
-                  child: Center(
-                    child: GestureDetector(
-                      onTap: _authenticateLocation,
-                      child: Image.asset(
-                        'assets/images/zarome.png',
-                        width: 250,
-                        height: 250,
-                        fit: BoxFit.contain,
-                        errorBuilder: (context, error, stackTrace) {
-                          return Container(
-                            width: 250,
-                            height: 250,
-                            color: Colors.grey[300],
-                            child: const Icon(
-                              Icons.person,
-                              size: 100,
-                              color: Colors.grey,
-                            ),
-                          );
-                        },
-                      ),
-                    ),
-                  ),
-                ),
-              ],
-            ),
-          ),
-          
-          const SizedBox(height: 50),
-          
-          Row(
-            mainAxisAlignment: MainAxisAlignment.center,
-            children: [
-              _buildStatCard(
-                icon: Icons.delete_outline,
-                value: "${_getActivityCount('trash')}개",
-                color: const Color(0xFF2E7D32),
-                onTap: () {},
-              ),
-              const SizedBox(width: 8),
-              _buildStatCard(
-                icon: Icons.restaurant_outlined,
-                value: "${_getActivityCount('food')}g",
-                color: const Color(0xFF2E7D32),
-                onTap: () {},
-              ),
-              const SizedBox(width: 8),
-              _buildStatCard(
-                icon: Icons.train_outlined,
-                value: "${_getActivityCount('transport')}회",
-                color: const Color(0xFF2E7D32),
-                onTap: () {},
-              ),
-              const SizedBox(width: 8),
-              _buildStatCard(
-                icon: Icons.recycling_outlined,
-                value: "${_getActivityCount('recycle')}개",
-                color: const Color(0xFF2E7D32),
-                onTap: () {},
-              ),
-            ],
-          ),
-          
-          const SizedBox(height: 30),
-          
-          Padding(
-            padding: const EdgeInsets.only(left: 20),
-            child: const Align(
-              alignment: Alignment.centerLeft,
-              child: Text(
-                '기록하기',
-                style: TextStyle(
-                  fontSize: 20,
-                  fontWeight: FontWeight.bold,
-                  color: Color(0xFF2E7D32),
-                ),
-              ),
-            ),
-          ),
-          const SizedBox(height: 20),
-          
-          Column(
-            children: [
-              Row(
-                mainAxisAlignment: MainAxisAlignment.center,
-                children: [
-                  _buildRecordCard(
-                    icon: Icons.delete_outline,
-                    title: '쓰레기줄이기',
-                    type: 'trash',
-                  ),
-                  const SizedBox(width: 20),
-                  _buildRecordCard(
-                    icon: Icons.restaurant_outlined,
-                    title: '잔반안남기기',
-                    type: 'food',
-                  ),
-                ],
-              ),
-              const SizedBox(height: 16),
-              Row(
-                mainAxisAlignment: MainAxisAlignment.center,
-                children: [
-                  _buildRecordCard(
-                    icon: Icons.train_outlined,
-                    title: '대중교통이용하기',
-                    type: 'transport',
-                  ),
-                  const SizedBox(width: 20),
-                  _buildRecordCard(
-                    icon: Icons.recycling_outlined,
-                    title: '분리수거하기',
-                    type: 'recycle',
-                  ),
-                ],
-              ),
-            ],
-          ),
-        ],
-      ),
-    );
-  }
-
-  Widget _buildRankingCard({
-    required String label,
-    required String value,
-  }) {
-    return Container(
-      padding: const EdgeInsets.symmetric(horizontal: 16, vertical: 12),
-      decoration: BoxDecoration(
-        color: const Color(0xFFE8F5E8),
-        borderRadius: BorderRadius.circular(20),
-        border: Border.all(
-          color: const Color(0xFF2E7D32).withOpacity(0.3),
-          width: 1,
-        ),
-      ),
-      child: Row(
-        mainAxisSize: MainAxisSize.min,
-        children: [
-          Text(
-            label,
-            style: const TextStyle(
-              fontSize: 14,
-              fontWeight: FontWeight.w600,
-              color: Color(0xFF2E7D32),
-            ),
-          ),
-          const SizedBox(width: 8),
-          Text(
-            value,
-            style: const TextStyle(
-              fontSize: 14,
-              fontWeight: FontWeight.bold,
-              color: Colors.black,
-            ),
-          ),
-          const SizedBox(width: 4),
-          const Icon(
-            Icons.trending_up,
-            size: 16,
-            color: Colors.red,
-          ),
-        ],
-      ),
-    );
-  }
-
-  int _getActivityCount(String type) {
-    try {
-      return _activityData.activityCounts[type] ?? 0;
-    } catch (e) {
-      return 0;
     }
   }
 
@@ -325,10 +39,8 @@ class _HomeSec1State extends State<HomeSec1> {
     Navigator.push(
       context,
       MaterialPageRoute(
-        builder: (context) => ActivityAuthPage(
-          activityType: type,
-          activityTitle: title,
-        ),
+        builder: (context) =>
+            ActivityAuthPage(activityType: type, activityTitle: title),
       ),
     ).then((result) {
       if (result != null && result is EcoActivity) {
@@ -339,115 +51,22 @@ class _HomeSec1State extends State<HomeSec1> {
     });
   }
 
-  Widget _buildRecordCard({
-    required IconData icon,
-    required String title,
-    required String type,
-  }) {
-    return GestureDetector(
-      onTap: () => _navigateToRecordPage(type, title),
-      child: Container(
-        width: 180,
-        height: 140,
-        decoration: BoxDecoration(
-          color: const Color(0xFFE8F5E8),
-          borderRadius: BorderRadius.circular(16),
-          border: Border.all(
-            color: const Color(0xFF2E7D32).withOpacity(0.3),
-            width: 1,
-          ),
-        ),
-        child: Padding(
-          padding: const EdgeInsets.all(30),
-          child: Column(
-            children: [
-              Expanded(
-                child: Row(
-                  mainAxisAlignment: MainAxisAlignment.spaceBetween,
-                  children: [
-                    Icon(
-                      icon,
-                      size: 48,
-                      color: const Color(0xFF2E7D32),
-                    ),
-                    Container(
-                      width: 40,
-                      height: 40,
-                      decoration: const BoxDecoration(
-                        color: Color(0xFF2E7D32),
-                        shape: BoxShape.circle,
-                      ),
-                      child: const Icon(
-                        Icons.add,
-                        color: Colors.white,
-                        size: 24,
-                      ),
-                    ),
-                  ],
-                ),
-              ),
-              Align(
-                alignment: Alignment.centerLeft,
-                child: Text(
-                  title,
-                  style: const TextStyle(
-                    fontSize: 14,
-                    fontWeight: FontWeight.w600,
-                    color: Color(0xFF2E7D32),
-                  ),
-                  maxLines: 1,
-                  overflow: TextOverflow.ellipsis,
-                ),
-              ),
-            ],
-          ),
-        ),
-      ),
-    );
-  }
+  @override
+  Widget build(BuildContext context) {
+    if (_isLoading) {
+      return const Center(child: CircularProgressIndicator());
+    }
 
-  Widget _buildStatCard({
-    required IconData icon,
-    required String value,
-    required Color color,
-    required VoidCallback onTap,
-  }) {
-    return GestureDetector(
-      onTap: onTap,
+    return SingleChildScrollView(
       child: Container(
-        width: 80,
-        height: 90,
-        decoration: BoxDecoration(
-          color: const Color(0xFFE8F5E8),
-          borderRadius: BorderRadius.circular(12),
-          border: Border.all(
-            color: color.withOpacity(0.3),
-            width: 1,
-          ),
-        ),
+        padding: const EdgeInsets.all(20),
         child: Column(
-          mainAxisAlignment: MainAxisAlignment.center,
           children: [
-            Icon(
-              icon,
-              size: 28,
-              color: color,
-            ),
-            const SizedBox(height: 8),
-            Container(
-              width: 30,
-              height: 1,
-              color: color,
-            ),
-            const SizedBox(height: 8),
-            Text(
-              value,
-              style: TextStyle(
-                fontSize: 13,
-                fontWeight: FontWeight.w600,
-                color: color,
-              ),
-            ),
+            const LocationAuthSection(),
+            const SizedBox(height: 50),
+            ActivityStatsSection(activityData: _activityData),
+            const SizedBox(height: 30),
+            ActivityRecordSection(onRecordTap: _navigateToRecordPage),
           ],
         ),
       ),
